+++ src/core/OS_utils.c
@@ -239,7 +239,7 @@ int test_available_space(int64_t req_size) {
 		return 1;
 	}
 	siril_debug_print("Tested free space ok: %ld for %ld MB free\n",
-			req_size / BYTES_IN_A_MB, free_space / BYTES_IN_A_MB);
+			(long)(req_size / BYTES_IN_A_MB), (long)(free_space / BYTES_IN_A_MB));
 	return 0;
 }
